+++ examples/OnDemandResources Example/OnDemandResourcesDemo/main.m
@@ -1,0 +1,19 @@
+//
+//  main.m
+//  OnDemandResourcesDemo
+//
+//  Created by xiejunyi on 2020/3/17.
+//  Copyright © 2020 xiejunyi. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
+
+int main(int argc, char * argv[]) {
+    NSString * appDelegateClassName;
+    @autoreleasepool {
+        // Setup code that might create autoreleased objects goes here.
+        appDelegateClassName = NSStringFromClass([AppDelegate class]);
+    }
+    return UIApplicationMain(argc, argv, nil, appDelegateClassName);
+}